requirements.txt
numpy
pandas
scikit-learn
torch
tqdm
openvino-dev
torchvision
albumentations
datasets
pillow
matplotlib
seaborn
timm
scikit-image
streamlit
Authlib